+++ CHANGELOG.md
@@ -973,7 +973,7 @@
 ### Fixes
 
 + Fixed Pocket Sack-In-A-Sack Replace in lore
-+ Fixed possible crash with broken neu repo when opening the desk inventory in garden (Contributed by CalMWolfs)
++ Fixed possible crash with broken neu repo when opening the desk inventory in the garden (Contributed by CalMWolfs)
 + Fixed frozen treasures per hour display being inaccurate (Contributed by CalMWolfs)
 + Fixed bug with ghost counter sometimes not detecting new kills (Contributed by CalMWolfs)
 + Fixed **Ghost Counter** - HiZe & ksipli
@@ -1286,7 +1286,7 @@
       and show a preview how these stats change when hovering over an upgrade
 + Hide crop money display, crop milestone display and garden visitor list while inside anita show, SkyMart or the
   composter inventory
-+ Hide chat messages from the visitors in garden. (Except Beth and Spaceman)
++ Hide chat messages from the visitors in the garden. (Except Beth and Spaceman)
 + Introduced a new command '/shcroptime <amount> <item>' that displays the estimated time it will take to gather the
   requested quantity of a particular item based on the current crop speed.
 + Show the average crop milestone in the crop milestone inventory.